+++ tools/xdhq/examples/ZND/Hello/main.php
@@ -52,7 +52,7 @@ $dom = new AtlasDOM();
 while ( true ) {
 	switch( $dom->getAction( $id ) ) {
 	case "Connect":
-		$dom->setLayout( "", new AtlasTree(), "Main.xsl" );
+		$dom->setLayout( "", file_get_contents( "h:/hg/epeios/tools/xdhq/examples/common/Hello/Main.html" ) );
 		break;
 	case "Typing":
 		$dom->setContent( "name", $dom->getContent( $id ) );