+++ tools/xdhq/wrappers/JRE/xdhqjre.cpp
@@ -20,7 +20,6 @@
 #include "xdhqjre.h"
 
 #include "registry.h"
-#include "treep.h"
 #include "xdhp.h"
 
 #include "iof.h"