+++ tools/xdhq/wrappers/ZND/xdhp.h
@@ -28,24 +28,7 @@ namespace xdhp {
 	SCLZND_F( Listen );
 	SCLZND_F( New );
 	SCLZND_F( GetAction );
-	SCLZND_F( Execute );
-	SCLZND_F( Alert );
-	SCLZND_F( Confirm );
-	SCLZND_F( SetLayout );
-	SCLZND_F( GetContents );
-	SCLZND_F( SetContents );
-	SCLZND_F( DressWidgets );
-	SCLZND_F( AddClasses );
-	SCLZND_F( RemoveClasses );
-	SCLZND_F( ToggleClasses );
-	SCLZND_F( EnableElements );
-	SCLZND_F( DisableElements );
-	SCLZND_F( SetAttribute );
-	SCLZND_F( GetAttribute );
-	SCLZND_F( RemoveAttribute );
-	SCLZND_F( SetProperty );
-	SCLZND_F( GetProperty );
-	SCLZND_F( Focus );
+	SCLZND_F( Launch );
 }
 
 #endif