+++ devel/CHANGELOG.md
@@ -5,6 +5,11 @@
 - [adpt]: Adaptation to changes in underlying modules;
 - [stdz]: modification to comply with current standards;
 
+
+## 2018-08-07
+- SCLZND :
+  - introducing 'sLong',
+
 ## 2018-08-07
 - STSFSM :
   - handling thread-safety with item rather then mutex.