+++ tools/zndq/CHANGELOG.md
@@ -1,5 +1,8 @@
 # *ZNDq* changelog
 
+## 2018-08-07
+- using `zend_long` (which can have a length of 32 or 64 bits) instead of `bso::s64`,
+
 ## 2018-04-28
 - handling the fact as, in *CGI* mode, the same script is called several times, but initialization and registering must only occur the first time,
 - handling the fact that *zndq* can be used simultaneously by different extension,